+++ pulsar/apps/test/result.py
@@ -1,5 +1,6 @@
 import sys
 import traceback
+from inspect import istraceback
 from copy import deepcopy
 
 
@@ -267,15 +268,18 @@ class TestResult(TestObject):
         """Converts a sys.exc_info()-style tuple of values into a string."""
         exctype, value, tb = err
         # Skip test runner traceback levels
-        while tb and self._is_relevant_tb_level(tb):
-            tb = tb.tb_next
-
-        if exctype is test.failureException:
-            # Skip assert*() traceback levels
-            length = self._count_relevant_tb_levels(tb)
-            msgLines = traceback.format_exception(exctype, value, tb, length)
+        if istraceback(tb):
+            while tb and self._is_relevant_tb_level(tb):
+                tb = tb.tb_next
+    
+            if exctype is test.failureException:
+                # Skip assert*() traceback levels
+                length = self._count_relevant_tb_levels(tb)
+                msgLines = traceback.format_exception(exctype, value, tb, length)
+            else:
+                msgLines = traceback.format_exception(exctype, value, tb)
         else:
-            msgLines = traceback.format_exception(exctype, value, tb)
+            msgLines = tb
 
         return ''.join(msgLines)