+++ pulsar/lib/__init__.py
@@ -1,7 +1,7 @@
 #First try local
 try:
     from ._pulsar import *
-    hasextensions = True
+    hasextensions = True    
 except ImportError:
     # Try Global
     try:
@@ -10,5 +10,9 @@ except ImportError:
     except ImportError:
         hasextensions = False
         from .fallback import *
+
+HTTP_REQUEST = 0
+HTTP_RESPONSE = 1
+HTTP_BOTH = 2
 
 from . import fallback